requirements.txt
torchaudio
torchcodec
torchvision
transformers==4.57.3
onnxruntime-gpu==1.23.2
tensorrt==10.14.1.48.post1
onnx==1.20.0
omegaconf
einops
gradio
